+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v3.22.0] - 2026-04-02
 ### Added
 - [#285] Creating auth-registration CRs for v2-dogus when LOP-IdP is used
 - [#285] Ignore auth-registration events when auth-registration is not enabled